ecommerce/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-crispy-forms==2.0
django-environ==0.10.0
django-pagination==1.0.7
django-simple-pagination==1.3
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
